Moved business logic to backend folder and intialised empty frontend folder
