+++ backend/configs/mailer.js
@@ -1,0 +1,11 @@
+const nodemailer = require('nodemailer');
+
+let transporter = nodemailer.createTransport({
+    host: 'smtp.gmail.com',
+    auth: {
+        user: process.env.MAIL_USERNAME,
+        pass: process.env.MAIL_PASSWORD
+    }
+});
+
+module.exports = transporter;
