+++ backend/service/cronJob.js
@@ -1,0 +1,8 @@
+const cron = require('node-cron');
+const { deleteData } = require('../repository/fileSharingRepository');
+
+cron.schedule('0 7 * * *', () => {
+    deleteData();
+});
+
+module.exports = cron;
