--- service/cronJob.js
@@ -1,8 +1,0 @@
-const cron = require('node-cron');
-const { deleteData } = require('../repository/fileSharingRepository');
-
-cron.schedule('0 7 * * *', () => {
-    deleteData();
-});
-
-module.exports = cron;
